+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Preserve image alternate text in DOCX ➡️ HTML/Markdown converter
 - Optionally write alternate text (if available) in place of images in DOCX ➡️ plain text converter
 - Support for grouped shapes and drawing canvas in DOCX ➡️ RTF converter
+- Fix: "pseudo inline" shapes created an incorrect layout in DOCX ➡️ RTF converter
 
 ## 0.18.1 - 2026.04.10